記憶度ボタンを押すと、次に復習するまでの日数が表示される
