+++ app/assets/stylesheets/question_answer/review.css
@@ -51,6 +51,27 @@
   border: 1px solid gray;
 }
 
+.review-option__notice-display-date {
+  width: 120px;
+  border: 1px solid gray;
+}
+
+.review-option__notice-display-date.js-0 {
+  color: blue;
+}
+
+.review-option__notice-display-date.js-1 {
+  color: blueviolet;
+}
+
+.review-option__notice-display-date.js-2 {
+  color: violet;
+}
+
+.review-option__notice-display-date.js-3 {
+  color: red;
+}
+
 .review-option__btn-wrap {
   display: flex;
   justify-content: space-around;
