+++ app/javascript/packs/application.js
@@ -14,7 +14,7 @@ require("../preview")
 require("../pulldown_font_image.js")
 require("../edit.js")
 require("../answer_appear.js")
-require("../repetition.js")
+require("../evaluate_memory_level.js")
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
