+++ app/views/question_answers/review.html.erb
@@ -14,6 +14,7 @@
           <div class="review-option__memory-level-description">
             <div class="review-option__memory-level-text">記憶度</div>
             <div class="hoge">?</div>
+            <div class="review-option__notice-display-date"></div>
           </div>
           <div class="review-option__btn-wrap">
             <div class="review-option__memory-level-low">低</div>
